+++ app/constants/FontSizes.js
@@ -1,0 +1,5 @@
+export default Object.freeze({
+  SMALL: 'fontSizeSmall',
+  MEDIUM: 'fontSizeMedium',
+  LARGE: 'fontSizeLarge',
+});