+++ src/domain/feature-flags/FeatureFlags.js
@@ -1,0 +1,3 @@
+export default Object.freeze({
+  FONT_SIZE_CONTROLS: 'fontSizeControls',
+});